+++ packages/polaris-viz/CHANGELOG.md
@@ -15,6 +15,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Added `arc` property to `PartialTheme` to allow override of the default theme option for `<DonutChart />`.
 - Added `legendFullWidth` prop to `<DonutChart />`
+- Added `renderInnerValueContent` prop to `<DonutChart />`.
 
 ## [7.11.1] - 2022-12-05